refactor(constants): Constants 수정
