+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/common/utils/Constants.kt
@@ -1,5 +1,5 @@
 package com.devfloor.hongit.api.common.utils
 
 const val BASE_API_URI = "/api"
-const val LOCAL_DATE_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ss"
+
 const val SEOUL_TIME_ZONE = "Asia/Seoul"
