+++ hongit-core/src/main/kotlin/com/devfloor/hongit/core/common/utils/Constants.kt
@@ -1,3 +1,3 @@
 package com.devfloor.hongit.core.common.utils
 
-const val LOCAL_DATE_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ss"
+const val LOCAL_DATE_TIME_FORMAT = "yyyy-MM-dd HH:mm:ss"
